@@ -56,27 +56,27 @@
     "express": "^4.17.1",
     "express-async-handler": "1.1.4",
     "express-force-ssl": "0.3.2",
-    "express-session": "1.16.1",
-    "flatted": "^2.0.0",
+    "express-session": "1.16.2",
+    "flatted": "^2.0.1",
     "forever-monitor": "1.7.1",
     "formidable": "1.2.1",
     "fs": "0.0.2",
     "http": "0.0.0",
-    "http-errors": "^1.7.2",
-    "http-shutdown": "^1.2.0",
+    "http-errors": "^1.7.3",
+    "http-shutdown": "^1.2.1",
     "httperrors": "2.3.0",
     "https": "1.0.0",
     "minimist": "^1.2.0",
-    "mongodb": "^3.2.6",
-    "mongoose": "^5.5.11",
+    "mongodb": "^3.2.7",
+    "mongoose": "^5.6.3",
     "mongoose-plugin-autoinc": "^1.1.9",
     "morgan": "^1.9.1",
-    "node": "^12.3.1",
+    "node": "^12.6.0",
     "on-finished": "2.3.0",
-    "openapi-backend": "^2.3.0",
+    "openapi-backend": "^2.3.4",
     "path": "0.12.7",
     "request": "^2.88.0",
-    "rotating-file-stream": "^1.4.1",
+    "rotating-file-stream": "^1.4.2",
     "rxjs": "^6.5.2",
     "serve-favicon": "2.5.0",
     "simple-get": "3.0.3",
@@ -144,17 +144,17 @@
     "sinon-chai": "3.3.0",
     "source-map-support": "^0.5.12",
     "style-loader": "^0.23.1",
-    "supports-color": "^6.1.0",
-    "ts-loader": "^6.0.1",
-    "ts-node": "^8.2.0",
-    "tslint": "^5.16.0",
+    "supports-color": "^7.0.0",
+    "ts-loader": "^6.0.4",
+    "ts-node": "^8.3.0",
+    "tslint": "^5.18.0",
     "tslint-eslint-rules": "^5.4.0",
     "tslint-microsoft-contrib": "^6.2.0",
     "tslint-no-unused-expression-chai": "^0.1.4",
-    "typescript": "^3.4.5",
-    "typescript-tslint-plugin": "^0.4.0",
-    "webpack": "^4.32.2",
-    "webpack-cli": "^3.3.2",
+    "typescript": "^3.5.2",
+    "typescript-tslint-plugin": "^0.5.3",
+    "webpack": "^4.35.2",
+    "webpack-cli": "^3.3.5",
     "why-is-node-running": "^2.1.0",
     "wtfnode": "^0.8.0"
   }
