@@ -1,10 +1,10 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, ErrorHandler } from '@angular/core';
 import { NGXLogger } from 'ngx-logger';
-import { map } from 'rxjs/operators';
+import { map, catchError } from 'rxjs/operators';
 import { Observable, of } from 'rxjs';
 
 import { MembersService } from '../../shared/members-service/members.service';
-import { IMember } from '../../api/api-members.service';
+import { IMember } from '../../data-providers/members.data-provider';
 
 /**
  * This component displays a dashboard showing key information on a number of members.
@@ -24,6 +24,7 @@ export class DashboardComponent implements OnInit {
   constructor(
     private membersService: MembersService,
     private logger: NGXLogger,
+    private errorHandler: ErrorHandler,
   ) {
     this.logger.trace(
       DashboardComponent.name + ': Starting DashboardComponent',
@@ -35,16 +36,28 @@ export class DashboardComponent implements OnInit {
   }
 
   /**
-   * Gets the members from the server by calling the membersService function.
+   * Gets the members from the server.
    */
   getMembers(): Observable<IMember[]> {
     this.logger.trace(DashboardComponent.name + ': Calling getMembers');
+
+    let errorHandlerCalled = false;
+
     return this.membersService.getMembers().pipe(
       map((members) => {
         return members.slice(
           this.firstMemberOnDisplay - 1,
           this.lastMemberOnDisplay,
         );
+      }),
+      catchError((error: any) => {
+        /* only call the error handler once per ngOnInit even though the returned observable is multicast to multiple html elements */
+        if (!errorHandlerCalled) {
+          errorHandlerCalled = true;
+          this.errorHandler.handleError(error);
+        }
+        /* return dummy value */
+        return of([]);
       }),
     );
   }
