@@ -1,15 +1,17 @@
-import { Component, OnInit, OnDestroy } from '@angular/core';
-import { Observable, Subject } from 'rxjs';
+import { Component, OnInit, ErrorHandler } from '@angular/core';
+import { Observable, Subject, of } from 'rxjs';
 import {
   debounceTime,
   distinctUntilChanged,
   switchMap,
-  takeUntil,
+  publishReplay,
+  refCount,
+  catchError,
 } from 'rxjs/operators';
 import { NGXLogger } from 'ngx-logger';
 
 import { MembersService } from '../../shared/members-service/members.service';
-import { IMember } from '../../api/api-members.service';
+import { IMember } from '../../data-providers/members.data-provider';
 import { config } from '../../config';
 
 /**
@@ -20,25 +22,28 @@ import { config } from '../../config';
   templateUrl: './member-search.component.html',
   styleUrls: ['./member-search.component.scss'],
 })
-export class MemberSearchComponent implements OnInit, OnDestroy {
+export class MemberSearchComponent implements OnInit {
   /* main title */
   header = 'Member Search';
+  /* initialises search hint */
+  isStart = true;
+  /* observable of array of members returned from search */
+  members$: Observable<IMember[]> = of([]);
+  /* subject observable to initiate search */
+  private searchTerms$ = new Subject<string>();
+  /* search debounce time in ms */
+  private debounce = 300;
   /* base route to get member detail */
   detail = config.routes.detail;
   /* member property to display in the list of found members */
   propertyToDisplay = 'name';
-  /* members stored locally from search term */
-  members: IMember[] | undefined;
-  /* subject observable to initiate search */
-  private searchTerms$ = new Subject<string>();
-  /* unsubscribe signal */
-  private unsubscribeSignal$ = new Subject<void>();
-  /* search debounce time in ms */
-  private debounce = 300;
+  /* controls that errorHandler only called once */
+  private errorHandlerCalled = false;
 
   constructor(
     private membersService: MembersService,
     private logger: NGXLogger,
+    private errorHandler: ErrorHandler,
   ) {
     this.logger.trace(
       MemberSearchComponent.name + ': Starting MemberSearchComponent',
@@ -49,28 +54,33 @@ export class MemberSearchComponent implements OnInit, OnDestroy {
     /**
      * Creates an observable of the members returned by the search term, subscribes and stores the found members.
      */
-    this.searchTerms$
-      .pipe(
-        /* wait a set interval after each keystroke before considering the term */
-        debounceTime(this.debounce),
+    this.members$ = this.searchTerms$.pipe(
+      /* wait a set interval after each keystroke before considering the term */
+      debounceTime(this.debounce),
 
-        /* ignore new term if same as previous term */
-        distinctUntilChanged(),
+      /* ignore new term if same as previous term */
+      distinctUntilChanged(),
 
-        /* get the memberService observable each time the term changes */
-        switchMap(
-          (term: string): Observable<IMember[]> => {
-            return this.membersService.getMembers(term);
-          },
-        ),
+      /* get the memberService observable each time the term changes */
+      switchMap(
+        (term: string): Observable<IMember[]> => {
+          return this.membersService.getMembers(term);
+        },
+      ),
+      /* using publish as share will resubscribe for each html call in case of unexpected error causing observable to complete (and I don't need to resubscribe on this page) */
+      publishReplay(1),
+      refCount(),
 
-        /* unsubscribe on ngDestroy */
-        takeUntil(this.unsubscribeSignal$.asObservable()),
-      )
-      .subscribe((foundMembers) => {
-        /* store returned members locally */
-        this.members = foundMembers;
-      });
+      catchError((error: any) => {
+        /* only call the error handler once per ngOnInit even though the returned observable is multicast to multiple html elements */
+        if (!this.errorHandlerCalled) {
+          this.errorHandlerCalled = true;
+          this.errorHandler.handleError(error);
+        }
+        /* return empty array of members to all html elements */
+        return of([]);
+      }),
+    );
   }
 
   /**
@@ -86,7 +96,8 @@ export class MemberSearchComponent implements OnInit, OnDestroy {
    */
   clear(): void {
     this.logger.trace(MemberSearchComponent.name + ': Calling clear()');
-    this.members = undefined;
+    /* getMembers('') returns an empty array without querying the backend */
+    this.searchTerms$.next('');
   }
 
   /**
@@ -98,10 +109,5 @@ export class MemberSearchComponent implements OnInit, OnDestroy {
 
   trackByFn(_index: number, member: IMember): number | null {
     return member ? member.id : null;
-  }
-
-  ngOnDestroy(): void {
-    this.unsubscribeSignal$.next();
-    this.unsubscribeSignal$.unsubscribe();
   }
 }
